Add fat/floppy.c
A volume boot record for FAT12 (or FAT16) floppies with a custom
boot protocol for the subsequent boot phases.

- 8086 compatible code. Defensive programming taken very seriously,
  e.g. it is assumed that BIOS service calls trash all registers.

- Sector sizes 512, 1024, 2048, 4096 are supported, although it is
  mostly theoretical to find sizes other than 512 or 1024. Boot code
  checks that BIOS Parameter Block contains a valid sector size.

- Most of the BIOS Parameter Block values are indirectly tested so
  that definitely invalid values are not used (mostly zeros).

- Boot sector code makes sure that it is executed at physical address
  0x00007C00. The exact register CS and IP values do not matter.

- Diskette parameter table is copied and values for "sector size" and
  "last sector on track" are patched.

- For disk access, INT 0x13, AH=0x02 is used. If it fails, there is
  a special pattern of "retry retry reset retry retry". A disk reset
  is not done after the first read fail.

- The correct video page is used if teletyping the error message.

The second stage loader is a file LOADER.512 in the FAT12/FAT16 root
directory and the file size is exactly 512 bytes. Boot sector code
does not read file allocation tables. The file can be anywhere in the
root directory and the directory is read a sector at a time until the
file is found. The second stage loader must have a correct checksum.
Boot sector code gives the read sector routine address to the second
stage loader.