Add fat/ldr512.c
Here is the second stage loader that is read and executed by all the
boot records provided here and it is a normal file, "LOADER.512", in
the root directory. The size of this file is exactly 512 bytes and it
implements the custom boot protocol used in conjunction with volume
boot records. For example, there is a shared procedure, read_sector,
that abstracts the differences between CHS and LBA.

The main function of this second stage loader is to be an extention
to the volume boot record and to read another file from the FAT file
system. This other file, "LOADER.AT", is the third stage loader and
the size of this file is exactly 32768 bytes. File allocation tables
are used and the file can be fragmented and be anywhere in the file
system. The format of "LOADER.AT" is not specified but the sum of all
words (2-byte little endian) must be 0x7C00.

From "VBR + LOADER.512" to "LOADER.AT" interface informally:

- "LOADER.AT" is loaded at memory location 0x00008000 and because its
  size is 32768 bytes, the memory range is from 0x8000 to 0xFFFF.

- CS:IP = 0x0000:0x8000, ES = SS = DS = 0x0000, SP = 0x7C00, other
  registers are undefined but the state of the computer in general is
  assumed to be the same as it were if BIOS started an MBR or VBR.
  (Please note the register DL, "boot drive", see below.)

- BIOS Parameter Block is at memory location 0x00007C00 and some of
  those values are patched, e.g. "sectors per track", "heads", and
  "drive number". The register DL given by BIOS can be found here.

The third stage loader is the "ultimate" boot loader and it is a
relatively extensive project of its own.