Add a template program
Parsing command-line arguments is relatively complicated if all the
common conventions and corner cases are taken into account. Here is
my attempt to solve the problem. The program is called "cmd" and it
takes a few options, option-arguments, and operands.

If starting the program with a long option "--help", it should print
the following output on stdout:

 > Usage: cmd [-a] [-b] [-m message]... [-o output] message...
 >
 > Options:
 >     -a  enable option a
 >     -b  enable option b
 >     -m  add header message
 >     -o  output file
 >
 > General:
 >     --help     help text
 >     --version  version information

For starting the program, here are some uncommon examples:

 > cmd -a -b Hello World
 > cmd -aaaabbbbaaaa -m "My header 1" -m "My Header 2" Hello World
 > cmd Hello World -bbaa
 > cmd Hello World -bbaa -- --this-is-not-an-option
 > cmd Hello World -otest.txt
 > cmd Hello World -o test.txt
 > cmd Hello World -o -

For error handling, these examples might be interesting:

 > cmd Hello World -z
 > cmd Hello -o
 > cmd World -m

Return values are EXIT_SUCCESS or EXIT_FAILURE. The idea is that the
main() handles command-line arguments and builds an option structure
that is given to the program(). Ideally these should be as loosely
coupled as possible but that does not happen in this version. Further
research might be needed and some details are subject to change.