Improve the template program
Usage instructions should use a better text layout. If there were an
unknown/invalid option, the error message would be printed before the
usage instructions. If starting the program with a "--not-an-option":

>Error: unknown long option "--not-an-option"
>
>Usage: cmd [-a] [-b] [-m message]... [-o output] message...
>
>Options:
>  -a            enable option a
>  -b            enable option b
>  -m message    add header message
>  -o output     output file
>
>General:
>  --help        help text
>  --version     version information

Return value is EXIT_FAILURE (for "--help" it would be EXIT_SUCCESS).

Another improvement is that the program() does not need to check and
handle argc == 0 which is very unlikely on systems these command line
tools are supposed to be executed because options, option-arguments,
and operands are needed for program input. The usage instructions are
still printed and the program returns EXIT_FAILURE.

In addition to this, the previous interface for program() has changed
because the program name, argv[0], is not an operand and might not be
very useful (as a concept in general) for portable programs.