Add dy-mbr.c and dy-vbr.c
Two portable programs for patching disk image files. These tools will
patch an existing master boot record or volume boot record. For the
former, the partition table area is not modified. For the latter, the
BIOS Parameter Block is not modified (bytes from 0x0003 to 0x003D).
There is a simple test for verifying that the file system layout is a
"standard" FAT12 or FAT16.

This commit also introduces a simple Makefile. Advanced users might
want to try this out on Linux. The following commands should work:

        make
        mkdosfs -I -C bin/hddtest.img 31744
        bin/dy-vbr bin/hddtest.img

This image should be bootable and have a Dancy volume boot record in
it that reads a file LOADER.512 from the root directory. This binary
file is in bin/LOADER.512 and can be copied into the file system.
However, step by step instructions for this are not provided because
the third stage loader, LOADER.AT, is not yet available.

This same can be applied to a partitioned media. Then bin/dy-mbr will
be used for the whole media and the bin/dy-vbr for the partition. In
the future, there will be native tools for creating file systems (and
partitions) so external tools, e.g. mkdosfs, are not used. This was
only for demonstrating purposes.