Improve the Makefile
This version of Makefile is still simple but has been revised. For a
full operating system, this kind of Makefile would not work but here
the basic principle is to keep things simple and introduce advanced
new features only when they are really needed.

Dancy source tree files have names that are compatible with a "8.3"
filename convention. There are exceptions, e.g. all .gitignore files,
but these are not part of the actual source tree. The version control
system is not tightly coupled with Dancy, although the commit history
and especially commit messages are part of the Dancy project. If this
version control system is changed, all these commit messages will be
extracted, converted into a proper format, and archived.