Improve dy-mbr.c and dy-vbr.c
There are two reasons for the improvements. First of all, the verbose
option was missing and for these tools it is now used for printing out
checksums of the binary blobs. The second reason, being important too,
was to check how relatively trivial changes like these behave when it
comes to diff logs and stability of the programs. At a first glance it
seems that the command line parsing method was very easy to extend to
support these new options.