Prepare for a new build system
There will be a rewritten build system introduced in the next commit
and this is just for preparing the current source code tree for this
relatively major modification. Diff logs for the next commit should
be much cleaner this way because there is no need to compare the new
Makefile to the previous version, i.e. changes are too fundamental.

In general, this is also an experiment of how to handle modifications
like these when it comes to version control systems. There might be a
lesson to be learned, so this method is not necessarily the best one,
and much less an established practice that shall be used from now on.