Introduce the new build system
There was much more work and research involved than I thought. It is
possible that some parts of the build system infrastructure need to
be redesigned when native object files are generated for the kernel.
Like with the previous Makefile implementation, I try to keep things
simple and do incremental changes as soon as they are needed. Now it
is possible to build the tools on wide range of different platforms
and "bootstrap" the Dancy Operating System. It should be clear that
the project is at an early stage and some of the fundamental design
choices are not made yet. However, a vision of the overall structure
has been in progress in some time already. To mention an example, it
is not necessary to reinvent all the ideas but to use knowledge from
other projects. The applications/implementations of the ideas may be
sometimes called "reinventing the wheel" but there is a significant
difference between these two aspects.

The design and implementation are equally important.