Add dy-blob.c
This tool is used for writing files LOADER.512 and LOADER.AT. With all
these things currently in the Dancy source tree, it should be possible
to create a bootable disk on a Linux-like operating system:

WARNING! Do not use these commands if you are not familiar with them.

  make
  dd if=/dev/zero of=floppy.img bs=512 count=2880
  losetup /dev/loop0 floppy.img
  mkdosfs /dev/loop0
  mount -t msdos /dev/loop0 /mnt/floppy
  cp LOADER.* /mnt/floppy/
  umount /mnt/floppy
  losetup -d /dev/loop0
  bin/dy-vbr floppy.img

I will write a tool that generates the file system and adds the files
into it. Those commands above are not, and never will be, part of the
building process. In fact, in a previous commit message I wrote:

 "In the future, there will be native tools for creating file systems
  (and partitions) so external tools, e.g. mkdosfs, are not used. This
  was only for demonstrating purposes."

These tools are quite straightforward to implement, and definitely can
be considered "reinventing the wheel", but the points are to keep the
toolchain as stand-alone as possible and to improve coding skills. The
programming "language" for the LOADER.AT was assembly, exclusively. It
is time to move on and change the way of thinking from assembly to C.