Add dy-init.c
This is an important step forward. In the past, I have done all tests
on LOADER.AT without having a good, "productized", tool for generating
file headers and checksums of the init executables. I did those with a
kind of assembler macro, and relying on ad-hoc tools written in C. The
current tool is not guaranteed to be bug-free but I'm quite happy with
the overall structure of it and the way this diff log shows the needed
steps when integrating a new tool into the Dancy source tree.

There is still some work to do before I do any "announcements" of this
project. At minimum, there should be bootable images available and a
public interface specification of the LOADER.AT boot services. I hope
it will be much easier to follow the progress when I have reached the
stage of having a system that can be booted and the infrastructure of
this source tree is more stable.