Add eltorito.asm
I guess I have a quite unique way of supporting bootable CDs. I just
use the ISO 9660 file system only for storing the FAT image that uses
2048-byte sectors. There are no hacks but it is really straighforward.
The ELTORITO.BIN chain-loads the first sector of the image file and
updates the "hidden sectors" value in the Bios Parameter Table. As far
as the VBR, LOADER.512, and LOADER.AT see it, it is just like a hard
disk drive with partitions. The difference is that now we use sectors
that are not 512 bytes but 2048 bytes. There are no data outside the
ISO 9660 file system structures. The ELTORITO.BIN and the file system
image are normal files in the root directory. No hard-coded offsets
or other tricks (other than what El Torito needs).

It was really nice to see that the whole boot chain supported sector
sizes other than 512 bytes but is was not a coincidence because it
was designed that way. Nevertheless, I was expecting that there were
some bugs that prevented the boot. Of course, it is not proven to be
bug-free (can it be?) but at least it worked. There will be a lot more
tests because the bootable CD might be the primary distribution media.

There is quite a lot of code that uses the sector size. All the VBRs,
LOADER.512, and LOADER.AT. All the FAT structures, directories, init
files and test databases were read from the file system and it really
seems that it handles different sector sizes correctly.