Introduce the legacy flag
Architecture init files, in_ia16.at, in_ia32.at, and in_x64.at, have
a header that is 32 bytes in size. There is a so-called Legacy Flag
that tells the boot loader whether the executable is ready for taking
care of processors that are not very modern. The in_ia16.at file does
not use this flag. The default setting is to keep it unset because it
simplifies the init files and rest of the booting process.

If this flag is unset (0), then for both the in_ia32.at and in_x64.at,
executable images can assume that the floating point unit is available
and ready for use, the CPUID instruction is available, and 4-MiB pages
can be used. This means that the CPU is equivalent to the "Pentium" or
better, in layman's terms.

For the in_x64.at, there are checks whether the LAHF/SAHF instruction
is available (64-bit mode) and can the CMPXCHG16B instruction be used.

I updated the dy-init program so that the legacy flag can be set by
using the "--legacy" option. Then the loader only checks if the 32-bit
protected mode is available or if the 64-bit long mode is available.
It is good to follow the "don't repeat yourself (DRY)" principle but
here I made an exception and copied the CRC-32C function to the same
source code file and the reason for this is to make this program more
stand-alone. If someone wants to take the Dancy boot loader and write
a "bootable" program, this should make it easier.