Introduce experimental key codes
Having an elegant key-code translation scheme has been a challenge in
my previous system-level programming experiments. I am trying to find
a solution to this problem but some further research is still needed.
This commit is not necessarily the right direction but should make it
easier to make experiments with the boot loader. There are so called
"boot services" provided by the boot loader API and one of them gives
simple keyboard input messages. With these key codes it is possible to
interpret those messages and introduce some interactivity to the init
executables.

Init executables run in protected mode or long mode (ring 3) and shall
not initialize hardware until the "exit boot loader services" function
is called. Then the boot loader removes itself and returns to ring 0
and the system is fully stand-alone. So called bootable programs might
not need to reach this stage but they could use boot services like the
console output, keyboard input, block I/O, and serial ports. It should
be mentioned that this stage is only an intermediate step when booting
a full modern OS that takes care of hardware initialization and so on.

The fundamental goal of the Dancy Operating System is to be a modern
operating system. The native kernels, that are not available yet, will
disable boot services at some point.