Renew gpt.asm
There is a principle called "You aren't gonna need it" that says we
should implement things when we actually need them. I implemented one
feature called "Trusted Platform Module (TPM)" in the previous master
boot record (GPT) but that was not really used or tested. Other parts
of the Dancy boot chain don't really support this so I made a design
decision to remove this feature.

I'm not really sure would that feature been useful even if thoroughly
supported and implemented. Having that could have given a false sense
of security. Another big reason was, however, that I did't have enough
expertise in that technology.

Now the master boot record is simplified and it is possible to chain-
load the bootable partitions in a traditional way. Pre-made HDD images
with GPT partitions are not easy because the size of the media must be
known beforehand (assuming conformance to specification). There are a
few other problems with GPT-partitioned hard disk drives if using BIOS
firmware to boot the system. This master boot record (GPT) should work
on most computers but it requires LBA disk access routines.