Add early linker files
This program does not have any meaningful functionality at this point.
It reads input files and saves them in dynamically allocated memory
areas. I thought about this, whether I should use a rather complicated
buffering system and use less memory. I ended up using this relatively
simple approach, i.e. all input files are simply read and the program
can just process them like "unsigned char" arrays. I think my object
files would not be big enough that this became an issue, even in the
future. The other design decision (perhaps more controversial) is to
use plain "int" for saving the file size. I do know what type "size_t"
is but I still chose the plain "int" for this purpose.

I try to be very strict when it comes to type correctness. If I handle
the files like an "unsigned char array[]", I think it is more elegant
to strictly use a plain int when doing array subscripting, e.g.

        array[1];
        array[10];
        array[my_int_variable];

This design decision has nothing to do with file sizes, and possible
limitations, when it comes to the Dancy operating system. This is only
applied for tools like this linker that are reading relatively small
object files. It is unlikely that with 32-bit ints this would ever be
a problem. However, the tool will also support the standard's minimum
"INT_MAX" size, 0x7FFF. This size is expected to be a problem but the
functionality is for small experiments on ancient compilers, if I can
find ones. The "real" things will be build on a platform that should
have enough memory and "ints" of size 32-bit or more. Of course, all
errors are handled if limitations become a problem.

The whole point is to be aware of the C programming language and its
guaranteed limits. My current boot code supports the 8086-compatible
instructions (even if support just means a "not supported" message).
I guess it is reasonable to apply this same approach to C programming
and hope this results in portable and correct programs, in general.