Improve the early linker
There is a calm phase as far as low-level programming is concerned and
I am trying to figure out what kind of "object file interface" should
be used in the Dancy operating system project. The vision of this has
been quite clear, i.e. Dancy tools will fully handle the linking phase
of the compilation process.

As an example, an assembler could produce the following object files:

        func1.obj
        func2.obj

Then, a C compiler produces:

        func3.obj
        func4.obj

Then, the Dancy linker consumes the object files:

        dy-link -o test.bin func1.obj func2.obj func3.obj func4.obj

Details may vary but this is the main vision. This commit introduces a
"--dump" option that can be used for inspecting object files. The main
functionality of this linker is far from ready at this point. I am not
very happy with this current implementation because there are so many
"magic numbers" in the code. This is not a good thing in general when
it comes to readability of a source code file. It might be that some
parts will be rewritten but this is not a strict guarantee.

Why it is not a strict guarantee? If I wrote code with that assumption
in mind, i.e. "this is just a temporary implementation", it would have
too much effect on the quality and motivation. I may retroactively say
that this must be replaced with better code but there is a significant
difference. I am not writing temporary code on purpose or by default.