Improve the early linker and fix a bug
This commit is very small but there was an annoying bug in the main.c
file. Big object files were not supported correctly and the bug was a
leftover of temporary fault injections when I tested the read logic. I
have to be more careful because the bug was embarrassing and it could
have ended up in the "production release" if the size "chunk" were any
bigger than the current value. Although, I am expecting that most of
the object files are relatively small and would not be hit by the bug.

The latter is not a good thing. Exactly the opposite. I am not worried
about bugs that are so "obvious and visible" that will be found when
using the program. The worst kind of bugs do not have symptomps of any
kind but they silently propagate and live in a program. At some point,
they make a big announcement of their existence.