Improve the coding style
There should be more comments because they make the code layout more
balanced and easier to read. Like paragraphs in a written text. There
should not be many new features needed for the "--dump" option. I am
still working on fully understanding the object files and what would
be the proper way to implement the main functionality of the linker.