Improve the early linker
The commit adds more requirements when it comes to object files. The
validate_obj() function is very important because it should make it
possible to have much less checks later but it is a drawback if code
is reviewed without taking this into account. In general, things are
looking pretty good. I understand most of the parts that I will need
to get started with object files and simple executables.