Introduce MinGW-w64 cross compiler
This is still very experimental but it should be possible to build a
cross compiler on Unix-like systems. The external script

        ./scripts/external/mingw.sh

can be used for building the tool chain (binutils + gcc). There will
be clang.sh implemented at some point. On platforms other than Unix-
like system, binary releases of clang will be used if they are able
to output the object file format supported by the Dancy linker.