Restructure the early linker
There was too much code in one source code file. I hope this commit is
another step in the right direction. The external behavior should not
change unless I introduced bugs in the process. The code base is much
more maintainable now and extensibility should be easier. This project
is like a pilot project when it comes to multiple source code files.