Add a dump-ext option
The previous commit was very important. After restructuring the linker
code base, it has been more appealing to introduce trivial but lengthy
blocks of code. The "--dump-ext" feature would have been noise if it
were in the program.c file. Currently there is a whole file for dumps
so it does not make the overall code structure more complicated. Also,
I am now trying a naming convention in which the non-static functions
are named according to the source file.

dump.c
        dump_ext()
        dump_obj()

main.c
        main()

program.c
        program()

validate.c
        validate_obj()

The first part of the name is exactly the same as the source file name
and this might be a good way to organize code. Time will show whether
or not this works in practice.