Add symbol.c
This is really a placeholder. I have been studying the linkers so not
much new code is available now. Init executables must have relocation
information available but I am not sure what would be the best format
for embedding that into the init executables. The run-time code should
not need to do any validation tests when applying the relocations.