Flush code changes
There was and is a little break as far as operating system development
is concerned. I had small changes that were not committed a month ago
so I now "flush" them. Real life has been busy but I will concentrate
on programming and finish the linker in the near future.