Improve the linker
Some important features are not implemented yet, e.g. a proper error
message if there are multiply defined external symbols. This weekend
has been very productive and almost all "difficult" parts are solved.

There will be a lot of fine adjustments and bug fixes. This is a bit
too complicated and prone to bugs. The biggest problem is the lack of
well-defined interfaces between different parts of the program.