Add size limits
The easiest way to hit the limits is to use large .bss areas. I don't
want this linker to be able to write an output file that does not pass
the validation test that is used for input files. Assuming that the
output is an object file and not a Dancy executable with its header.