Change static symbol names
I changed the symbol name format mainly for aesthetic reasons but also
because there was a bug. If there were too many symbols, the sprintf
function would have written the terminating zero outside the buffer.

Also, there is a known bug in code that handles duplicate sections. I
will try to find it and fix it for the next commit.