Detect mangled names
The external symbols should have names that are compatible with the C
programming language. In some previous commit message I said that this
is not a strict rule. I would not be surprised if I had to allow these
mangled names for compatibility reasons in the future. At the moment
I think that "C identifiers ought to be enough for everyone", but the
Dancy operating system project is still in the beginning stages.