Improve the linker
Now the linker supports grouped sections. This was more complicated
than what I wanted it to be. If I had taken this requirement into
account from the very beginning, the solution could have been more
elegant. This commit has not been tested very well yet, but I think
that there should not be major "design" bugs. A lot of testing will
be done with this linker implementation and changes as big as this
should not be needed anymore. Striving for stability.