Introduce the native header
If the dy-link program uses the "-f at" option, the native header will
be added. These are like init executables but there is no relocation
code. The loader handles the everything, e.g. finding suitable memory
areas, sets page permissions (executable, writable, etc.) and so on.

I have not designed everything at this point. Should the executables
support dynamically linked libraries is an open question. For now, I
think that having those object-file-like executables gives a lot of
possibilities to try different things. All the information for full
relocation is always available. It adds some overhead for the loader
but perhaps on modern hardware that is not very significant.