Introduce the boot loader interface
This is not a well-written specification. The dancy.h header file have
prototypes of boot loader functions and a couple of enums and a struct
for video information. This commit also refactors header files. In the
future there might be a few changes to datatypes defined in the header
dancy/limits.h. I am aware of the C99 standard support for fixed width
integer types but currently those are not used. The kernel codebase is
not established yet but the whole source code structure is evolving.