Introduce dy-path test program
Perhaps this tool is an overkill for the intended purpose. I was a bit
uncertain whether the PATH variable set by Makefile really was usable.
It seems that it works like it should. This program could be used as a
template for new simple tools. Parsing the command line is by far the
most complicated part of this program.