Include stdarg header
This freestanding header is important. Having variadic functions in a
portable way is almost impossible without this header. The most known
function that uses this is printf. The kernel will have a version of
printf but it will not be called printf. One common name is kprintf,
i.e. "kernel printf" or something similar.