Use stricter alignment
One compiler produced an object file that has a section with 32-byte
alignment and the init files supported only 16-byte alignment. Now I
modified the program and the section alignment is 64 bytes for init
executables. In general, this is not a problem for other executables
because sections are not used "in-place".