Add initial version of dy-mcopy
The functionality of this program is not implemented yet. The command
line interface is taken from Mtools that is an open source collection
of tools to allow a system to modify files on FAT file systems. Only
one command (with a subset of Mtools features) will be implemented at
this point. The command is mcopy, just like the program name suggests.
The tool will automatically create a directory and support setting a
timestamp of the file. Example:

        dy-mcopy -i fdd1440.img system/IN_IA32.AT ::SYSTEM/IN_IA32.AT

There will be no advanced features, like long file names. It might be
that the directory hierarchy can have two levels only.