Add code for reading files
File system images that are supported by the boot loader are not very
big and that is one of the reasons why files are simply read wholly in
memory. The other reason is simplicity but there is still enough room
for bugs. I hope that I implemented the logic correctly.