Develop dy-mcopy
This commit introduces some important functionality but the program is
not ready. It took some time to parse the destination path name and I
tried to make it as robust as possible. The directory names and file
names use a stricter format than what the file system itself allows.

        dy-mcopy -i fdd160.img antti.txt ::my/long/path/antti.txt

        D: "MY         "
        D: "LONG       "
        D: "PATH       "
        F: "ANTTI   TXT"

There are a lot of error messages, e.g.

        dy-mcopy -i fdd160.img antti.txt ::toolongpath/antti.txt

        Error: length of "TOOLONGPATH"
        Error: incorrect path/destination

As always, I hope that there are no bugs. This program will be used
quite extensively so I would assume that major bugs will be detected.