Develop dy-mcopy
The main functionality is implemented but not tested. This project was
very stressful and exhausting because I wrote it too fast. I will not
integrate this with the release build scripts until I have extensively
tested the program. Fortunately there are many "check" tools on other
operating systems that can help verifying the file system structures.