Reach a milestone
The floppy image seems to boot. Everything else is just implementation
details and possible bug fixes (I assume that there are no major bugs,
but I cannot say for sure). The commit is a milestone because from now
on there will always be disk images available. There will be ISO 9660
images (CD-ROM) available in the future. Writing that tool is not the
top priority because floppy and USB disk images are enough for now.