Simplify build scripts
In the future, it is possible to write a tool that generates contents
for the system.mk. I am wondering whether it takes too much effort to
edit the build scripts manually whenever a new file is added into the
source tree. It could work so that there was a target in the Makefile
that searches the whole kernel directory and updates the system.mk. I
have to check if there is a de facto standard for that kind of target.