Renew build scripts
I refactored all the build scripts and this was a major modification.
It might not be extremely stable immediately but I do think the whole
system is a lot better now. There could be automatic tools that modify
files dirs.mk and objects.mk in the future. The format of those files
is very well suited for automatic processing.