Delete the template program
This tool was very good but it is not up-to-date anymore. The simplest
"template" program could be dy-path at the moment. It can be used when
starting to write new programs. If there were any bugs in command line
parsing, it should be fixed in every program. In general, I want that
those "dy" tools are a coherent collection, i.e. the same coding style
and potentially even the bugs (will be fixed simultaneously if found).