Modify other.mk
There were so many warnings in the host operating system headers. The
other system (unix.mk) seems to work much more sensibly when it comes
to compiler warnings. In general, I want to write code that generates
no warnings but only if warnings itself are correctable. For example

    Warning C4996: 'fopen': This function or variable may be unsafe.

I can't "fix" the warning because the "dy" tools are "C89" compatible.
Please note that this applies to tools only (that are used on the host
operating system). The native kernel code does not use this compiler.