Add string.c
There are only strcmp and strncmp functions available. At some point
I probably have to write a full C library but for now it is enough to
have the most important library functions. The init or kernel code do
not need (and should not have) the full library. It is the user space
that needs it but the system has not evolved to that point yet.

What I am really missing is a printf-like function. I will implemented
that (a suitable version for systems programming) very soon so that I
can print, e.g. the physical memory map (provided by the boot loader).