Introduce dy-iso program
The Dancy source tree is now a little bit more "self-hosting" because
a bootable .ISO is written by a "native" tool. If everything works as
intended, on a Unix-like computer it is possible to build the project:

        git clone https://github.com/Tiihala/Dancy.git
        cd Dancy
        make

It requires that the host system has "nasm" (an assembler) installed
and that the system is capable of building a cross compiler. However,
operating system hobbyists most probably have all prerequisite tools
available by default. The biggest drawback is that building the cross
compiler takes time but this is needed only once. A target "clean" in
the Makefile does not remove the external tools but "distclean" does.

        (No root privileges are needed, and should not be used)

The releases, disk images and the .ISO image, will be:

        release/dancy.iso
        release/fdd720.img
        release/fdd1440.img

I think now the general structure of this project has been defined.