Fix conversion warnings
If this code base were in production use, this kind of commit would be
dangerous. It might be that suppressing warnings cause more bugs. In
this case the Dancy source tree is still relatively "unstable" so this
kind of commit should be justified. Most of the changes are so trivial
that I am quite confident that this only improves the source code.