Add legacy in_ia16
The Dancy operating system should run on very old hardware. The modern
functionality is not available, of course. Whether this implementation
for legacy hardware should be called an "OS" is questionable. However,
it is like a client for the Dancy operating system instances that are
run on modern hardware (32-bit or 64-bit). Early 32-bit systems, like
80386, will also use this 16-bit version because modern versions have
certain requirements that the CPU must implement. The boot loader will
detect those CPU features and the appropriate version of the init, the
IN_IA16.AT, IN_IA32.AT, or IN_X64.AT.