Add initial version of dy-zip
The boot loader loads so called database files into memory. I did not
define the exact format of those but only the 32-byte header. The ZIP
file format can be used as a database file format, but certain little
modifications are needed. The most interesting feature of this format
is splitting the file into small parts, or "spanning multiple disks"
as they talked in the early nineties when the floppy disks were still
relevant for ordinary computer users.

I studied the file format and found out one small detail that causes
a problem. It seems that I cannot use my custom file header in files
that "continue the data from a previous archive". For example,

        example.z01
        example.z02     <--
        example.zip

The first file is not a problem. The last one could be a problem but
can most probably be solved. The .z02 cannot have my file header (if
I understood the file format correctly) when the compressed stream of
bytes is crossing the file boundary.

I try to come up with a good solution that solves the problem.