Develop dy-zip
The program can create a zip file but it is not compressed. There will
be special code for handling the Dancy database files. I have used zip
files for "all my life" and it feels very good to understand the basic
structure of this format. The extensions may make it complex but this
traditional version is relatively simple.