Develop dy-zip
This program is a bit too complex. It seems to work but in general it
is not easy to verify the correctness of the code. If the output file
name is "db_000.at", it will use the format that is supported by the
boot loader. Next I am going to study the Deflate algorithm. It might
take a while but fortunately there are many reference implementations.