Introduce bitarray functions
These function implementations are very trivial but it still took some
time to get the interfaces defined. This is actually a good example of
difficulties when it comes to the first part of the phrase "design and
implementation". I am not saying the interfaces are particularly good
but I will try use these and see how they work.