Fix bitarray_shove function
There was an enormous bug that made the whole function totally useless
and, once again, should have been found before releasing anything. I
really need to start writing unit tests because these kind of bugs are
not acceptable.

    The terms "quality assurance" and "quality control" are often
    used interchangeably to refer to ways of ensuring the quality
    of a service or product.

        - Wikipedia

Let's use the term "quality assurance" in this project. I will really
try to be more careful when releasing new source code. This particular
source code file is not thoroughly tested but I'm quite confident that
it does exactly what it was supposed to do.