Refactor common code
I am not very confident that source files should spread too much. The
reason for this refactoring is that tools and some kernel procedures
use the same source files. Everything in the common directory must be
very portable and conform to the limits defined by the C programming
language standard. The most obvious example is the "sizeof(int)". The
files in the kernel directory do not need to be that strict, although
the limits should be kept in mind in general when programming.