Introduce Huffman coding scheme
The implementation works with canonical Huffman codes. The source code
is written entirely from scratch but it uses some algorithms that are
not invented here. Thank you, Mark Adler (https://zlib.net).

I am very happy to understand the canonical Huffman codes and the way
codebooks can be transmitted. It is only necessary to send the number
of bits for each symbol, and symbols are just unsigned int values. The
Wikipedia has an example:

        A = 11          (bits 2)
        B = 0           (bits 1)
        C = 101         (bits 3)
        D = 100         (bits 3)

        unsigned symbols[] = { 2, 1, 3, 3 };

        huffman_init(&h, &symbols[0], 4);
        huffman_fetch(&h, &b);
        ...
        huffman_fetch(&h, &b);
        ...

The above example does not show the bitarray initialization. It should
be noted that "symbols" is not "const" because the same memory is used
for storing the actual symbol values.